+++ src/RcppExports.cpp
@@ -7,8 +7,8 @@
 using namespace Rcpp;
 
 // cpp_lda
-List cpp_lda(arma::sp_mat& mt, int k, int max_iter, double alpha, double beta, arma::sp_mat& seeds, arma::sp_mat& words, int random, int parallel, bool verbose);
-RcppExport SEXP _seededlda_cpp_lda(SEXP mtSEXP, SEXP kSEXP, SEXP max_iterSEXP, SEXP alphaSEXP, SEXP betaSEXP, SEXP seedsSEXP, SEXP wordsSEXP, SEXP randomSEXP, SEXP parallelSEXP, SEXP verboseSEXP) {
+List cpp_lda(arma::sp_mat& mt, int k, int max_iter, double alpha, double beta, arma::sp_mat& seeds, arma::sp_mat& words, int random, bool verbose);
+RcppExport SEXP _seededlda_cpp_lda(SEXP mtSEXP, SEXP kSEXP, SEXP max_iterSEXP, SEXP alphaSEXP, SEXP betaSEXP, SEXP seedsSEXP, SEXP wordsSEXP, SEXP randomSEXP, SEXP verboseSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -20,9 +20,8 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< arma::sp_mat& >::type seeds(seedsSEXP);
     Rcpp::traits::input_parameter< arma::sp_mat& >::type words(wordsSEXP);
     Rcpp::traits::input_parameter< int >::type random(randomSEXP);
-    Rcpp::traits::input_parameter< int >::type parallel(parallelSEXP);
     Rcpp::traits::input_parameter< bool >::type verbose(verboseSEXP);
-    rcpp_result_gen = Rcpp::wrap(cpp_lda(mt, k, max_iter, alpha, beta, seeds, words, random, parallel, verbose));
+    rcpp_result_gen = Rcpp::wrap(cpp_lda(mt, k, max_iter, alpha, beta, seeds, words, random, verbose));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -47,7 +46,7 @@ END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_seededlda_cpp_lda", (DL_FUNC) &_seededlda_cpp_lda, 10},
+    {"_seededlda_cpp_lda", (DL_FUNC) &_seededlda_cpp_lda, 9},
     {"_seededlda_cpp_lda0", (DL_FUNC) &_seededlda_cpp_lda0, 9},
     {NULL, NULL, 0}
 };