+++ tests/testthat/test-textmodel_seededlda.R
@@ -1,4 +1,3 @@
-
 require(quanteda)
 data(data_corpus_moviereviews, package = "quanteda.textmodels")
 
@@ -8,7 +7,7 @@ toks <- tokens(data_corpus_moviereviews[1:500],
                remove_number = TRUE)
 dfmt <- dfm(toks) %>%
     dfm_remove(stopwords(), min_nchar = 2) %>%
-    dfm_trim(max_docfreq = 0.1, docfreq_type = "prop")
+    dfm_trim(max_docfreq = 0.5, docfreq_type = "prop")
 sifi <- c("space", "mars", "alien", "earth")
 
 test_that("seeded LDA is working", {
@@ -18,7 +17,7 @@ test_that("seeded LDA is working", {
 
     set.seed(1234)
     lda <- textmodel_seededlda(dfmt, dict, residual = TRUE,
-                               min_termfreq = 100)
+                               min_termfreq = 5)
 
     expect_equal(dim(terms(lda, 10)), c(10, 3))
     expect_equal(dim(terms(lda, 20)), c(20, 3))
@@ -91,21 +90,28 @@ test_that("seeded LDA is working", {
 
 test_that("predict works with seeded LDA", {
 
-    dict <- dictionary(list(romance = c("love*", "couple*"),
-                            sifi = c("alien*", "star", "space")))
+    dict <- dictionary(list(romance = c("lover", "couple", "marige"),
+                            sifi = c("aliens", "star", "space")))
 
     dfmt_train <- head(dfmt, 450)
     dfmt_test <- tail(dfmt, 50)
 
     lda <- textmodel_seededlda(dfmt_train, dict, residual = TRUE)
-
+    #lda_old <- textmodel_seededlda(dfmt_train, dict, residual = TRUE, old = TRUE)
     pred_train <- predict(lda)
+    #pred_train_old <- predict(lda, old = TRUE)
     expect_equal(names(pred_train), docnames(dfmt_train))
     expect_equal(
         levels(pred_train),
         c("romance", "sifi", "other")
     )
     expect_true(sum(topics(lda) == pred_train) / length(pred_train) > 0.9)
+
+    head(topics(lda), 10)
+    head(unname(pred_train), 10)
+
+    head(topics(lda_old), 10)
+    head(unname(pred_train_old), 10)
 
     pred_test <- predict(lda, newdata = dfmt_test)
     expect_equal(names(pred_test), docnames(dfmt_test))